[sensorfw] Use setDataRate() for settings data rates. Fixes JB#60417
Sensorfwd setInterval() method takes time in milliseconds -parameter
which is reciprocal of requested data rate in Hertz. As the time value
is an integer, this severely limits the available data rates especially
in the high end to: 1000, 500, 333, 250, 200, ... Hz

Switch to using setDataRate() which is available in sensorfwd >= 0.13.0.
It is both conceptually better match with qt side sensor interfaces and
more future proof as data rate is transferred over D-Bus as floating
point value.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>